Changed the return type of getLevel of the Player

+++ src/main/java/unibo/exiled/controller/GameController.java
@@ -43,9 +43,9 @@ public interface GameController {
     /**
      * Gets the level of the player.
      *
-     * @return A double representing the level of the player.
+     * @return A integer representing the level of the player.
      */
-    double getPlayerLevel();
+    int getPlayerLevel();
 
     /**
      * Gets the class name of the player.
